use extension files for representer patching
